# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Donation Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (Admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (Admin only — all users must be created by MHQ admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a specific user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: User not found
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: User not found
  /users/{id}/toggle-active:
    patch:
      operationId: toggleUserActive
      tags: [users]
      summary: Activate or deactivate a user (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "404":
          description: User not found
  /users/{id}/reset-password:
    post:
      operationId: resetUserPassword
      tags: [users]
      summary: Reset a user's password (Admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordReset"
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: User not found

  # ─── EYE CALLS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── UNIT ─────────────────────────────────────────────────────────────────
...
    AppUser:
      type: object
      required: [id, clerkId, email, name, role, isActive, mustChangePassword, createdAt, updatedAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [super_admin, eye_bank_head, unit_coordinator]
        unitId:
          type: ["integer", "null"]
        unitName:
          type: ["string", "null"]
        isActive:
          type: boolean
        mustChangePassword:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UserInput:
      type: object
      required: [email, name, role]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          minLength: 1
        role:
          type: string
          enum: [super_admin, eye_bank_head, unit_coordinator]
        unitId:
          type: ["integer", "null"]
        password:
          type: string
          minLength: 8
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [super_admin, eye_bank_head, unit_coordinator]
        unitId:
          type: ["integer", "null"]
        isActive:
          type: boolean
...
    PasswordReset:
      type: object
      required: [newPassword]
      properties:
        newPassword:
          type: string
          minLength: 8
...